Update DLL and APM00300

+++ Bimasakti-version/1.00/PROGRAM/BS Program/SOURCE/COMMON/GL/GLM00400COMMON/DTOs/GLM00414/GLM00415DTO.cs
@@ -11,7 +11,7 @@
         public string CYEAR { get; set; }
         public string CCENTER_CODE { get; set; }
         public string CCENTER_NAME { get; set; }
-        public int NVALUE { get; set; }
+        public decimal NVALUE { get; set; }
     }
